+++ taxel-gui/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod report_store;
+
 use anyhow::{Context, Result};
 use log::debug;
 use std::{